Refactor playerControls Hook
